requirements.txt
ImageHash~=4.0
numpy~=1.18.1
Pillow~=7.0.0
scikit-learn~=0.22.2.post1
tensorflow~=2.1.0
tqdm~=4.42.1
